Added a big number formatter
UI that displays groing numbers can now reference BigNumberFormatter in order to shorten numbers. EX 1000 will now read as 1k or 1 Thousand. amount of decimals can also be adjust. EX 12345 can be 12k or 12.345k. This works up to a decillion or d. After that, numbers will be displayed with scientific notation. EX 1000 Decillion will be 1E33 (or something idk thats a lot of zeros)