Added Cheats
We're getting to the point where we can balance the game, and we need to tone down earnings. I've added cheats so we can still test features that will appear deeper into a run. Press ` to activate. The full list of commands is in the Google Doc under "Unity Editing." The current script is much too complicated. Cleanup has been added to ToDo, but it makes the feature functional for now.